+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Additions
 - Added the ability for Dropdown to display an arrow texture based on its open state
 - Added the ability to specify Dropdown paragraph colors through style properties
 - Added the UiLayouter class and ILayoutItem interface to allow using MLEM.Ui's layouting system for custom ui systems
+- Added the ability to set a selected texture and color for buttons
 
 Improvements
 - Explicitly return the element type from Dropdown.AddElement overloads